+++ registry/components/ui/resizable/index.ts
@@ -1,0 +1,3 @@
+export { default as ResizablePaneGroup } from "./resizable-pane-group.svelte";
+export { default as ResizablePane } from "./resizable-pane.svelte";
+export { default as ResizableHandle } from "./resizable-handle.svelte";